Represent colors same as ruby SASS does.
All colors are now represented as either rgba(int, int, int, float) or
as #rrggbb. hsl and hsla are converted to these formats as well.

The impetus for this fix was that colors were sometimes represented with
percents, but on zeros those percent signs would be stripped, causing
invalid formats:

rgb(51.765%, 0%, 0.566%)

was converted to

rgb(51.765%, 0, 0.566%)

which is invalid. (All arguments must either be integers or percents; 0
is interpreted by the browser as an integer)

Unknown if there are other builtin functions that fall prey to this
percent-stripping bug as well, but this gets us partway there and
results in better output anyway.

(As an aside, there was special code preserving float arguments to rgb,
which wasn't useful because CSS defines the colorspace to be 32 bits.)